requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
groq==0.11.0
python-dotenv==1.0.1
